Fix Warnings in p3gxscircles
/libretroshare/src/services/p3gxscircles.cc:1005: warning:
Cppcheck(stlSize): Possible inefficient checking for 'groupIds' emptiness.
/libretroshare/src/services/p3gxscircles.cc:43: warning: zero as null
pointer constant
/libretroshare/src/services/p3gxscircles.cc:182: warning: zero as null
pointer constant
/libretroshare/src/services/p3gxscircles.cc:645: warning: zero as null
pointer constant

/libretroshare/src/services/p3gxscircles.cc:968: warning: implicit
conversion changes signedness: 'int' to 'uint32_t' (aka 'unsigned int')
/libretroshare/src/services/p3gxscircles.cc:1283: warning: zero as null
pointer constant
/libretroshare/src/services/p3gxscircles.cc:1617: warning: use of old-
style cast
/libretroshare/src/services/p3gxscircles.cc:1636: warning:
'getTokenService' is deprecated
/libretroshare/src/services/p3gxscircles.cc:1655: warning:
'getTokenService' is deprecated
/libretroshare/src/services/p3gxscircles.cc:1701: warning: implicit
conversion changes signedness: 'int' to 'uint32_t' (aka 'unsigned int')
/libretroshare/src/services/p3gxscircles.cc:1733: warning: implicit
conversion changes signedness: 'unsigned int' to 'int'
/libretroshare/src/services/p3gxscircles.cc:1733: warning: implicit
conversion changes signedness: 'int' to 'unsigned int'
/libretroshare/src/services/p3gxscircles.cc:1746: warning: implicit
conversion changes signedness: 'unsigned long' to 'int'
/libretroshare/src/services/p3gxscircles.cc:1719: warning: implicit
conversion loses integer precision: 'std::__cxx11::list::size_type' (aka
'unsigned long') to 'int'
/libretroshare/src/services/p3gxscircles.cc:1990: warning: zero as null
pointer constant
/libretroshare/src/services/p3gxscircles.cc:1992: warning: implicit
conversion loses integer precision: 'uint32_t' (aka 'unsigned int') to
'uint8_t' (aka 'unsigned char')
/libretroshare/src/services/p3gxscircles.cc:1990: warning: implicit
conversion loses integer precision: 'time_t' (aka 'long') to 'uint32_t'
(aka 'unsigned int')
/libretroshare/src/services/p3gxscircles.cc:1997: warning: implicit
conversion loses integer precision: 'size_t' (aka 'unsigned long') to
'uint32_t' (aka 'unsigned int')
/libretroshare/src/services/p3gxscircles.cc:1998: warning: implicit
conversion loses integer precision: 'size_t' (aka 'unsigned long') to
'uint32_t' (aka 'unsigned int')
/libretroshare/src/services/p3gxscircles.cc:2002: warning: implicit
conversion loses integer precision: 'size_t' (aka 'unsigned long') to
'uint32_t' (aka 'unsigned int')
/libretroshare/src/services/p3gxscircles.cc:2088: warning: zero as null
pointer constant
/libretroshare/src/services/p3gxscircles.cc:2124: warning: zero as null
pointer constant

+++ libretroshare/src/services/p3gxscircles.h
@@ -212,7 +212,7 @@ virtual RsServiceInfo getServiceInfo();
 
 	protected:
 
-	bool pushCircleMembershipRequest(const RsGxsId& own_gxsid,const RsGxsCircleId& circle_id,uint32_t request_type) ;
+	bool pushCircleMembershipRequest(const RsGxsId& own_gxsid, const RsGxsCircleId& circle_id, uint8_t request_type) ;
 	static uint32_t circleAuthenPolicy();
 
 	/** Notifications **/
